requirements.txt
Babel==2.2.0
Django==1.8.11
django-admin-bootstrapped==2.5.7
django-ckeditor==5.0.3
django-phonenumber-field==1.0.0
django-recaptcha2==0.1.7
django-storages==1.4
furl==0.4.92
orderedmultidict==0.7.5
phonenumbers==7.2.7
Pillow==3.1.1
psycopg2==2.6.1
pytz==2016.2
requests==2.9.1
six==1.10.0
